Synchronized with WebKit @ r219671 (Safari 10.1.2 / Safari-603.3.8)

+++ Source/WebCore/crypto/parameters/CryptoAlgorithmHmacKeyParams.h
@@ -37,7 +37,7 @@ public:
     // FIXME: Consider merging hash and hashIdentifier.
     JSC::JSValue hash;
     CryptoAlgorithmIdentifier hashIdentifier;
-    std::optional<size_t> length;
+    std::optional<unsigned long> length;
 
     Class parametersClass() const final { return Class::HmacKeyParams; }
 };
